Began work on separating the unit tests from the integration tests. Moved TestingTool.cs to main project.

+++ Webpay.Integration.CSharp/Webpay.Integration.CSharp/Order/Row/OrderRowBuilder.cs
@@ -105,7 +105,7 @@
         /// </summary>
         /// <param name="quantity"></param>
         /// <returns>OrderRowBuilder</returns>
-        public OrderRowBuilder SetQuantity(int quantity)
+        public OrderRowBuilder SetQuantity(decimal quantity)
         {
             _quantity = quantity;
             return this;
